各種fbへの対応
