+++ src/components/Header/Header.module.scss
@@ -37,7 +37,7 @@
         }
     }
     
-    .h1 {
+    h1 {
         font-family: 'Inter';
         font-style: normal;
         font-weight: 700;
