+++ src/components/Index/AboutFushigiFes/AboutFushigiFes.module.scss
@@ -16,19 +16,19 @@
         font-weight: bold;
         margin-bottom: 10px;
     }
+    h2 {
+        text-align: center;
+        font-weight: 700;
+        font-size: 24px;
+        line-height: 29px;
+    }
+    
 }
 
 .inner {
     width: 95%;
     max-width: 1200px;
     margin: 0 auto 30px;
-}
-
-.h2 {
-    text-align: center;
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 29px;
 }
 
 .textAlignCenter {
