+++ src/components/Index/AboutFushigiFes/AboutFushigiFes.tsx
@@ -1,11 +1,11 @@
 import styles from '@/components/Index/AboutFushigiFes/AboutFushigiFes.module.scss'
-import GreenButton from '@/components/Button/GreenButton'
+import GreenButtonLink from '@/components/ButtonLink/GreenButtonLink'
 
 const AboutFushigiFes = () => {
   return (
     <div className={styles.aboutFushigiFes} id="about-fushigi-fes">
     <div className={styles.inner}>
-        <h2 className={styles.h2}>不思議RTAフェスとは？</h2>
+        <h2>不思議RTAフェスとは？</h2>
         <div className={styles.textContent}>
             不思議のダンジョンRTAフェス(略称:不思議RTAフェス)は、「不思議のダンジョンシリーズ」及び「その他のローグライクゲーム」をテーマにしたイベント開催などを行っている団体です。<br />
             <strong>RTA</strong>という遊び方を通して本ジャンルの面白さを伝えるための活動を主に行っています。<br /><br />
@@ -20,7 +20,7 @@ const AboutFushigiFes = () => {
             長年愛され続けているローグライクゲームのさらなる発展を願い、より多くの方に楽しさを知ってもらうことを目標に活動しています。
         </div>
         <div className={styles.textAlignCenter}>
-            <GreenButton text="もっと長ければ個別ぺーじ" href="#"/>
+            <GreenButtonLink text="もっと長ければ個別ぺーじ" href="#"/>
         </div>
     </div>
 </div>
