+++ src/components/Index/ComunityLinks/ComunityLinks.module.scss
@@ -7,6 +7,12 @@
 
 .comunityLink {
     padding: 60px 0;
+    h2 {
+        text-align: center;
+        font-weight: 700;
+        font-size: 24px;
+        line-height: 29px;
+    }
 }
 
 .comunityLink .comunityLinkFlex {
@@ -20,12 +26,6 @@
     }
 }
 
-.h2 {
-    text-align: center;
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 29px;
-}
 @media (max-width : 768px) {
     .comunityLink .comunityLinkFlex {
         flex-direction: column;
