+++ src/components/Index/EventArchives/EventArchives.module.scss
@@ -3,6 +3,15 @@
     padding: 70px;
     text-align: center;
     background: #EEEEEE;
+
+    h2 {
+        text-align: center;
+        font-weight: 700;
+        font-size: 24px;
+        line-height: 29px;
+        padding-bottom: 60px;
+    }
+
     ul {
         display: flex;
         max-width: 980px;
@@ -45,14 +54,6 @@
         text-decoration: none;
     }        
 
-}
-
-.h2 {
-    text-align: center;
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 29px;
-    padding-bottom: 60px;
 }
 
 @media (max-width : 768px) {
