+++ src/components/Index/SpecialThanks/SpecialThanks.module.scss
@@ -23,14 +23,14 @@
         }
     }
     
+    h2 {
+        text-align: center;
+        font-weight: 700;
+        font-size: 24px;
+        line-height: 29px;
+    }
 }
 
-.h2 {
-    text-align: center;
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 29px;
-}
 
 @media (max-width : 768px) {
